Moved get_teams to new competetions v4
